pyproject.toml
[project]
name = "llm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.5.3",
    "langchain>=0.3.28",
    "langchain-anthropic>=0.3.22",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.2.2",
    "truststore>=0.10.4",
]
